BUG: GHS3DPlugin: Fixed a bug in the max memory computation
The default maximum memory was defined as a short integer which caused the maximum memory to be badly initialised on computers with big memory.
If the available memory was more than 32768 Mo (2**15 Mo) the max memory was cast from int to short and could be initialised with any value included a negative one. In this case a 20Mo default value was used by GHS3D

+++ src/GHS3DPlugin/GHS3DPlugin_Hypothesis.cxx
@@ -101,7 +101,7 @@ bool GHS3DPlugin_Hypothesis::GetToMeshHoles(bool checkFreeOption) const
 //function : SetMaximumMemory
 //=======================================================================
 
-void GHS3DPlugin_Hypothesis::SetMaximumMemory(short MB)
+void GHS3DPlugin_Hypothesis::SetMaximumMemory(int MB)
 {
   if ( myMaximumMemory != MB ) {
     myMaximumMemory = MB;
@@ -114,7 +114,7 @@ void GHS3DPlugin_Hypothesis::SetMaximumMemory(short MB)
 //           * automatic memory adjustment mode. Default is zero
 //=======================================================================
 
-short GHS3DPlugin_Hypothesis::GetMaximumMemory() const
+int GHS3DPlugin_Hypothesis::GetMaximumMemory() const
 {
   return myMaximumMemory;
 }
@@ -123,7 +123,7 @@ short GHS3DPlugin_Hypothesis::GetMaximumMemory() const
 //function : SetInitialMemory
 //=======================================================================
 
-void GHS3DPlugin_Hypothesis::SetInitialMemory(short MB)
+void GHS3DPlugin_Hypothesis::SetInitialMemory(int MB)
 {
   if ( myInitialMemory != MB ) {
     myInitialMemory = MB;
@@ -135,7 +135,7 @@ void GHS3DPlugin_Hypothesis::SetInitialMemory(short MB)
 //function : GetInitialMemory
 //=======================================================================
 
-short GHS3DPlugin_Hypothesis::GetInitialMemory() const
+int GHS3DPlugin_Hypothesis::GetInitialMemory() const
 {
   return myInitialMemory;
 }
@@ -766,7 +766,7 @@ bool GHS3DPlugin_Hypothesis::DefaultMeshHoles()
 #include <windows.h>
 #endif
 
-short  GHS3DPlugin_Hypothesis::DefaultMaximumMemory()
+int  GHS3DPlugin_Hypothesis::DefaultMaximumMemory()
 {
 #ifndef WIN32
   struct sysinfo si;
@@ -795,7 +795,7 @@ short  GHS3DPlugin_Hypothesis::DefaultMaximumMemory()
 //function : DefaultInitialMemory
 //=======================================================================
 
-short  GHS3DPlugin_Hypothesis::DefaultInitialMemory()
+int  GHS3DPlugin_Hypothesis::DefaultInitialMemory()
 {
   return DefaultMaximumMemory();
 }
@@ -1368,7 +1368,7 @@ std::string GHS3DPlugin_Hypothesis::CommandToRun(const GHS3DPlugin_Hypothesis* h
   // Default memory is defined at ghs3d installation but it may be not enough,
   // so allow to use about all available memory
   if ( m ) {
-    short aMaximumMemory = hyp ? hyp->myMaximumMemory : -1;
+    int aMaximumMemory = hyp ? hyp->myMaximumMemory : -1;
     cmd += " -m ";
     if ( aMaximumMemory < 0 )
       cmd += DefaultMaximumMemory();
@@ -1376,7 +1376,7 @@ std::string GHS3DPlugin_Hypothesis::CommandToRun(const GHS3DPlugin_Hypothesis* h
       cmd += aMaximumMemory;
   }
   if ( M && !useBndRecovery ) {
-    short aInitialMemory = hyp ? hyp->myInitialMemory : -1;
+    int aInitialMemory = hyp ? hyp->myInitialMemory : -1;
     cmd += " -M ";
     if ( aInitialMemory > 0 )
       cmd += aInitialMemory;
