+++ src/GHS3DPlugin/GHS3DPlugin_Hypothesis_i.hxx
@@ -58,14 +58,14 @@ class GHS3DPLUGIN_EXPORT GHS3DPlugin_Hypothesis_i:
   /*!
    * Maximal size of memory to be used by the algorithm (in Megabytes)
    */
-  void SetMaximumMemory(CORBA::Short MB) throw ( SALOME::SALOME_Exception );
-  CORBA::Short GetMaximumMemory();
+  void SetMaximumMemory(CORBA::Long MB) throw ( SALOME::SALOME_Exception );
+  CORBA::Long GetMaximumMemory();
   /*!
    * Initial size of memory to be used by the algorithm (in Megabytes) in
    * automatic memory adjustment mode. Default is zero
    */
-  void SetInitialMemory(CORBA::Short MB) throw ( SALOME::SALOME_Exception );
-  CORBA::Short GetInitialMemory();
+  void SetInitialMemory(CORBA::Long MB) throw ( SALOME::SALOME_Exception );
+  CORBA::Long GetInitialMemory();
   /*!
    * Optimization level: 0-none, 1-light, 2-medium, 3-strong. Default is medium
    */
